+++ app/helpers/comments_helper.rb
@@ -1,0 +1,5 @@
+module CommentsHelper
+  def comments_formatted_for_csv(comments)
+    comments.pluck(:body).map(&:strip).join("|")
+  end
+end