+++ app/services/budget/export.rb
@@ -1,11 +1,14 @@
 class Budget
   class Export
+    include CommentsHelper
+
     HEADERS = [
       "RODA identifier",
       "Partner organisation identifier",
       "Partner organisation",
       "Level",
-      "Title"
+      "Title",
+      "Level B activity comments"
     ]
 
     def initialize(source_fund:, organisation: nil)
@@ -52,7 +55,8 @@ class Budget
         activity.partner_organisation_identifier,
         activity.extending_organisation&.name,
         I18n.t("table.body.activity.level.#{activity.level}"),
-        activity.title
+        activity.title,
+        activity.programme? ? comments_formatted_for_csv(activity.comments) : ""
       ]
     end