+++ spec/features/beis_users_can_download_exports_spec.rb
@@ -1,6 +1,8 @@
 RSpec.feature "BEIS users can download exports" do
+  let(:beis_user) { create(:beis_user) }
+
   before do
-    authenticate! user: create(:beis_user)
+    authenticate! user: beis_user
   end
   after { logout }
 
@@ -147,12 +149,17 @@ RSpec.feature "BEIS users can download exports" do
 
     report = create(:report)
 
+    programme = create(:programme_activity, extending_organisation: partner_organisation)
     project = create(:project_activity, :newton_funded, extending_organisation: partner_organisation)
+
+    create(:budget, financial_year: 2021, value: 1000, parent_activity: programme, report: nil)
 
     create(:budget, financial_year: 2018, value: 100, parent_activity: project, report: report)
     create(:budget, financial_year: 2019, value: 80, parent_activity: project, report: report)
     create(:budget, financial_year: 2020, value: 75, parent_activity: project, report: report)
     create(:budget, financial_year: 2021, value: 20, parent_activity: project, report: report)
+
+    programme_comment = create(:comment, commentable: programme, owner: beis_user, body: "I am a programme comment")
 
     visit exports_path
     click_link partner_organisation.name
@@ -160,15 +167,28 @@ RSpec.feature "BEIS users can download exports" do
 
     document = CSV.parse(page.body.delete_prefix("\uFEFF"), headers: true).map(&:to_h)
 
-    expect(document.size).to eq(1)
+    expect(document.size).to eq(2)
 
     expect(document).to match_array([
+      {
+        "RODA identifier" => programme.roda_identifier,
+        "Partner organisation identifier" => programme.partner_organisation_identifier,
+        "Partner organisation" => programme.extending_organisation.name,
+        "Level" => "Programme (level B)",
+        "Title" => programme.title,
+        "Level B activity comments" => programme_comment.body,
+        "2018-2019" => "0.00",
+        "2019-2020" => "0.00",
+        "2020-2021" => "0.00",
+        "2021-2022" => "1000.00"
+      },
       {
         "RODA identifier" => project.roda_identifier,
         "Partner organisation identifier" => project.partner_organisation_identifier,
         "Partner organisation" => partner_organisation.name,
         "Level" => "Project (level C)",
         "Title" => project.title,
+        "Level B activity comments" => "",
         "2018-2019" => "100.00",
         "2019-2020" => "80.00",
         "2020-2021" => "75.00",
@@ -186,6 +206,8 @@ RSpec.feature "BEIS users can download exports" do
     project1 = create(:project_activity, :newton_funded, extending_organisation: partner_organisation1, parent: programme)
     project2 = create(:project_activity, :newton_funded, extending_organisation: partner_organisation2, parent: programme)
 
+    create(:budget, financial_year: 2021, value: 1000, parent_activity: programme, report: nil)
+
     create(:budget, financial_year: 2018, value: 100, parent_activity: project1, report: report)
     create(:budget, financial_year: 2019, value: 80, parent_activity: project1, report: report)
     create(:budget, financial_year: 2021, value: 20, parent_activity: project1, report: report)
@@ -196,20 +218,36 @@ RSpec.feature "BEIS users can download exports" do
     create(:budget, financial_year: 2021, value: 20, parent_activity: project2, report: report)
     create(:budget, financial_year: 2021, value: 60, parent_activity: project2, report: report)
 
+    programme_comment_1 = create(:comment, commentable: programme, owner: beis_user, body: "I like big budgets and I cannot lie")
+    programme_comment_2 = create(:comment, commentable: programme, owner: beis_user, body: "The chief budgerigar asked if we could budge over this budget")
+
     visit exports_path
     click_link "Download Budgets for Newton Fund"
 
     document = CSV.parse(page.body.delete_prefix("\uFEFF"), headers: true).map(&:to_h)
 
-    expect(document.size).to eq(3)
+    expect(document.size).to eq(4)
 
     expect(document).to match_array([
+      {
+        "RODA identifier" => programme.roda_identifier,
+        "Partner organisation identifier" => programme.partner_organisation_identifier,
+        "Partner organisation" => programme.extending_organisation.name,
+        "Level" => "Programme (level B)",
+        "Title" => programme.title,
+        "Level B activity comments" => [programme_comment_1, programme_comment_2].map(&:body).join("|"),
+        "2018-2019" => "0.00",
+        "2019-2020" => "0.00",
+        "2020-2021" => "0.00",
+        "2021-2022" => "1000.00"
+      },
       {
         "RODA identifier" => project1.roda_identifier,
         "Partner organisation identifier" => project1.partner_organisation_identifier,
         "Partner organisation" => partner_organisation1.name,
         "Level" => "Project (level C)",
         "Title" => project1.title,
+        "Level B activity comments" => "",
         "2018-2019" => "100.00",
         "2019-2020" => "80.00",
         "2020-2021" => "0.00",
@@ -221,6 +259,7 @@ RSpec.feature "BEIS users can download exports" do
         "Partner organisation" => partner_organisation2.name,
         "Level" => "Project (level C)",
         "Title" => project2.title,
+        "Level B activity comments" => "",
         "2018-2019" => "100.00",
         "2019-2020" => "80.00",
         "2020-2021" => "75.00",
@@ -232,6 +271,7 @@ RSpec.feature "BEIS users can download exports" do
         "Partner organisation" => partner_organisation2.name,
         "Level" => "Project (level C)",
         "Title" => project2.title,
+        "Level B activity comments" => "",
         "2018-2019" => "0.00",
         "2019-2020" => "0.00",
         "2020-2021" => "0.00",