+++ CHANGELOG.md
@@ -1117,6 +1117,8 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 - Configure Rollout and Rollout UI gems to allow BEIS users to manage feature flags
 - Allow comments on Level B activities from BEIS users
 - Add comments column to Level B activities bulk upload template
+- Add Level B activity comments to budget exports
+- Strip leading/trailing whitespace and line breaks from comments in reports and exports
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-120...HEAD
 [release-120]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-119...release-120